requirements.txt
anyio==3.6.2
blinker==1.6.2
certifi==2023.5.7
click==8.1.3
Flask==2.3.2
h11==0.14.0
httpcore==0.17.1
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
python-telegram-bot==20.3
sniffio==1.3.0
starlette==0.27.0
uvicorn==0.22.0
Werkzeug==2.3.4
